move update helper to sg_update
