+++ dev/sg/sg_update.go
@@ -7,11 +7,15 @@ import (
 	"os"
 	"runtime"
 	"strings"
+	"time"
 
 	"github.com/urfave/cli/v2"
 
+	"github.com/sourcegraph/sourcegraph/dev/sg/internal/analytics"
 	"github.com/sourcegraph/sourcegraph/dev/sg/internal/download"
+	"github.com/sourcegraph/sourcegraph/dev/sg/internal/run"
 	"github.com/sourcegraph/sourcegraph/dev/sg/internal/std"
+	"github.com/sourcegraph/sourcegraph/dev/sg/root"
 	"github.com/sourcegraph/sourcegraph/lib/errors"
 	"github.com/sourcegraph/sourcegraph/lib/output"
 )
@@ -65,4 +69,67 @@ func updateToPrebuiltSG(ctx context.Context) (string, error) {
 		return "", err
 	}
 	return currentExecPath, nil
+}
+
+func checkSgVersionAndUpdate(ctx context.Context, out *std.Output, skipUpdate bool) error {
+	start := time.Now()
+
+	if BuildCommit == "dev" {
+		// If `sg` was built with a dirty `./dev/sg` directory it's a dev build
+		// and we don't need to display this message.
+		out.Verbose("Skipping update check on dev build")
+		return nil
+	}
+
+	_, err := root.RepositoryRoot()
+	if err != nil {
+		// Ignore the error, because we only want to check the version if we're
+		// in sourcegraph/sourcegraph
+		return nil
+	}
+
+	rev := strings.TrimPrefix(BuildCommit, "dev-")
+	revOut, err := run.GitCmd("rev-list", fmt.Sprintf("%s..origin/main", rev), "--", "./dev/sg")
+	if err != nil {
+		if strings.Contains(revOut, "bad revision") {
+			// installed revision is not available locally, that is fine - we wait for the
+			// user to eventually do a fetch
+			return errors.New("current sg version not found - you may want to run 'git fetch origin main'.")
+		}
+
+		// Unexpected error occured
+		analytics.LogEvent(ctx, "auto_update", []string{"check_error"}, start)
+		return err
+	}
+
+	revOut = strings.TrimSpace(revOut)
+	if revOut == "" {
+		// No newer commits found. sg is up to date.
+		return nil
+	}
+
+	if skipUpdate {
+		out.WriteLine(output.Styled(output.StyleSearchMatch, "╭──────────────────────────────────────────────────────────────────╮  "))
+		out.WriteLine(output.Styled(output.StyleSearchMatch, "│                                                                  │░░"))
+		out.WriteLine(output.Styled(output.StyleSearchMatch, "│ HEY! New version of sg available. Run 'sg update' to install it. │░░"))
+		out.WriteLine(output.Styled(output.StyleSearchMatch, "│       To see what's new, run 'sg version changelog -next'.       │░░"))
+		out.WriteLine(output.Styled(output.StyleSearchMatch, "│                                                                  │░░"))
+		out.WriteLine(output.Styled(output.StyleSearchMatch, "╰──────────────────────────────────────────────────────────────────╯░░"))
+		out.WriteLine(output.Styled(output.StyleSearchMatch, "  ░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░"))
+
+		analytics.LogEvent(ctx, "auto_update", []string{"skipped"}, start)
+		return nil
+	}
+
+	out.WriteLine(output.Line(output.EmojiInfo, output.StyleSuggestion, "Auto updating sg ..."))
+	if _, err := updateToPrebuiltSG(ctx); err != nil {
+		analytics.LogEvent(ctx, "auto_update", []string{"failed"}, start)
+		return errors.Newf("failed to install update: %s", err)
+	}
+	out.WriteSuccessf("sg has been updated!")
+	out.Write("To see what's new, run 'sg version changelog'.")
+
+	analytics.LogEvent(ctx, "auto_update", []string{"updated"}, start)
+
+	return nil
 }
